setSerial: do not use default device type
When setting the serial number with the wrong device type the SD-Wire
device becomes unusable until the serial number is set again specifying the
correct device type. So using 'sd-mux' as default device type is unwise.

Let's require that a device type is specified when executing the set serial
command.

Signed-off-by: Heinrich Schuchardt <xypron.glpk@gmx.de>